will try to do more sockets on char creation
